ipq807x: add 512MB memory profile
Most cheaper IPQ807x devices shipped with
only 512MB of RAM so they need to use a reduced memory profile in ath11k
as well as reduced reserved-memory for Q6 in order to free up some RAM.

So add a DTSI that overrides the default 1GB reserved-memory nodes with
the 512MB version.

Note that the ath11k profile is a config option that must be manually
selected.

Signed-off-by: Robert Marko <robimarko@gmail.com>